libgit2: refactor managed transport feature usage in reconciler and checkout pkg
Adds a new field `Managed` to `CheckoutOpts` to let checkout impls
figure out when to use managed transport. Adds a new field to the
reconciler which stores the state of managed transport registration and
uses the same to return a stalling error when managed transports are not
registered but the feature is enabled.

Signed-off-by: Sanskar Jaiswal <jaiswalsanskar078@gmail.com>